Quantum block related functionality in a separate commit
